+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Ingest from dataframe
 - Dependency issues
 - Typing issues
+- CloudInfo / CloudSettings now cached by authority (schema, host and port) instead of full URL
 
 ## [5.0.2] - 2025-05-30